# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AppUserRole": {
        "description": "`app_user.role`.",
        "enum": [
          "SALES",
          "ADMIN"
        ],
        "title": "AppUserRole",
        "type": "string"
      },
      "AppUserStatus": {
        "description": "`app_user.status`.",
        "enum": [
          "ACTIVE",
          "DISABLED"
        ],
        "title": "AppUserStatus",
        "type": "string"
      },
...
      "AuthenticatedUser": {
        "additionalProperties": false,
        "description": "[`AuthenticatedUser`](/architecture/interfaces.md#authenticateduser).",
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/AppUserRole"
          }
        },
        "required": [
          "id",
          "email",
          "display_name",
          "role"
        ],
        "title": "AuthenticatedUser",
        "type": "object"
      },
...
      "DemoLoginRequest": {
        "additionalProperties": false,
        "description": "[`DemoLoginRequest`](/architecture/interfaces.md#demologinrequest).",
        "properties": {
          "role": {
            "$ref": "#/components/schemas/AppUserRole"
          }
        },
        "required": [
          "role"
        ],
        "title": "DemoLoginRequest",
        "type": "object"
      },
      "Dependency": {
        "description": "`details.dependency` of `UPSTREAM_UNAVAILABLE`, exactly the\n[Dependencies](/architecture/interfaces.md#conventions) table's column: a wire-only enum,\nowned by interfaces and defined here once.",
        "enum": [
          "DATABASE",
          "CLASSIFIER",
          "LLM",
          "EMBEDDER",
          "HUBSPOT"
        ],
        "title": "Dependency",
        "type": "string"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "description": "`error` of [Conventions](/architecture/interfaces.md#conventions) `ErrorEnvelope`.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "$ref": "#/components/schemas/ErrorDetails",
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorDetailField": {
        "additionalProperties": false,
        "description": "One entry of `details.fields[]` ([Conventions](/architecture/interfaces.md#conventions)\n`VALIDATION`).",
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "ErrorDetailField",
        "type": "object"
      },
      "ErrorDetails": {
        "additionalProperties": false,
        "description": "`details` of [Conventions](/architecture/interfaces.md#conventions) `ErrorEnvelope`: the\nclosed set of names the Envelope table gives. Every field is optional and, per the Naming\nparagraph (G10), absent rather than `null` when unset.",
        "properties": {
          "dependency": {
            "$ref": "#/components/schemas/Dependency"
          },
          "entity_id": {
            "format": "uuid",
            "title": "Entity Id",
            "type": "string"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetailField"
            },
            "title": "Fields",
            "type": "array"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "resets_at": {
            "format": "date-time",
            "title": "Resets At",
            "type": "string"
          },
          "retry_after_min": {
            "title": "Retry After Min",
            "type": "integer"
          }
        },
        "title": "ErrorDetails",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "The `{\"error\": {...}}` shape of [Conventions](/architecture/interfaces.md#conventions),\nnamed `ErrorEnvelope` in the OpenAPI document.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "LoginRequest": {
        "additionalProperties": false,
        "description": "[`LoginRequest`](/architecture/interfaces.md#loginrequest).",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "password": {
            "format": "password",
            "title": "Password",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "User": {
        "additionalProperties": false,
        "description": "[`User`](/architecture/interfaces.md#user).",
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_login_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "role": {
            "$ref": "#/components/schemas/AppUserRole"
          },
          "status": {
            "$ref": "#/components/schemas/AppUserStatus"
          }
        },
        "required": [
          "id",
          "email",
          "display_name",
          "role",
          "status",
          "last_login_at"
        ],
        "title": "User",
        "type": "object"
      },
      "UserCreate": {
        "additionalProperties": false,
        "description": "[`UserCreate`](/architecture/interfaces.md#usercreate). `password`'s `PASSWORD_MIN_LENGTH`\nis enforced once, by `auth.users.create_user` (its one owner), not a static `Field`\nconstraint: the minimum is a runtime setting, and a model built inside a router factory\ncannot be resolved by FastAPI's OpenAPI generation (`from __future__ import annotations`\nturns the route's own annotation into a forward reference that must be a module-level\nname).",
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "password": {
            "format": "password",
            "title": "Password",
            "type": "string",
            "writeOnly": true
          },
          "role": {
            "$ref": "#/components/schemas/AppUserRole"
          }
        },
        "required": [
          "email",
          "display_name",
          "role",
          "password"
        ],
        "title": "UserCreate",
        "type": "object"
      },
      "UserUpdate": {
        "additionalProperties": false,
        "description": "[`UserUpdate`](/architecture/interfaces.md#userupdate). `password`'s `PASSWORD_MIN_LENGTH`\nis enforced once, by `auth.users.update_user` (`UserCreate`'s note).",
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "password": {
            "anyOf": [
              {
                "format": "password",
                "type": "string",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AppUserRole"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AppUserStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "UserUpdate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "LeadRadar API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/demo-login": {
      "post": {
        "description": "`API-78`.",
        "operationId": "demo_login_api_v1_auth_demo_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedUser"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Demo Login",
        "tags": [
          "authentication-and-users"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedUser"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Login",
        "tags": [
          "authentication-and-users"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout_api_v1_auth_logout_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "leadradar_session",
            "required": true,
            "schema": {
              "title": "Leadradar Session",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Logout",
        "tags": [
          "authentication-and-users"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "me_api_v1_auth_me_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "leadradar_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Leadradar Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedUser"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Me",
        "tags": [
          "authentication-and-users"
        ]
      }
    },
...
    "/api/v1/users": {
      "get": {
        "operationId": "get_users_api_v1_users_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "leadradar_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Leadradar Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/User"
                  },
                  "title": "Response Get Users Api V1 Users Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Get Users",
        "tags": [
          "authentication-and-users"
        ]
      },
      "post": {
        "operationId": "create_user_route_api_v1_users_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "leadradar_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Leadradar Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Create User Route",
        "tags": [
          "authentication-and-users"
        ]
      }
    },
    "/api/v1/users/{user_id}": {
      "patch": {
        "operationId": "update_user_route_api_v1_users__user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "leadradar_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Leadradar Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Update User Route",
        "tags": [
          "authentication-and-users"
        ]
      }
    }
  }
}